+++ tv.go
@@ -214,10 +214,7 @@ type TVWatchProvidersAppend struct {
 // Supports append_to_response.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-details
-func (c *Client) GetTVDetails(
-	id int,
-	urlOptions map[string]string,
-) (*TVDetails, error) {
+func (c *Client) GetTVDetails(id int64, urlOptions map[string]string) (*TVDetails, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d?api_key=%s%s",
@@ -251,11 +248,7 @@ type TVAccountStates struct {
 // If it belongs to your favourite list.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-account-states
-//
-func (c *Client) GetTVAccountStates(
-	id int,
-	urlOptions map[string]string,
-) (*TVAccountStates, error) {
+func (c *Client) GetTVAccountStates(id int64, urlOptions map[string]string) (*TVAccountStates, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/account_states?api_key=%s%s",
@@ -314,10 +307,7 @@ type TVAggregateCredits struct {
 // GetTVAggregateCredits get the aggregate credits (cast and crew) that have been added to a TV show.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-aggregate-credits
-func (c *Client) GetTVAggregateCredits(
-	id int,
-	urlOptions map[string]string,
-) (*TVAggregateCredits, error) {
+func (c *Client) GetTVAggregateCredits(id int64, urlOptions map[string]string) (*TVAggregateCredits, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/aggregate_credits?api_key=%s%s",
@@ -343,10 +333,7 @@ type TVAlternativeTitles struct {
 // GetTVAlternativeTitles get all of the alternative titles for a TV show.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-alternative-titles
-func (c *Client) GetTVAlternativeTitles(
-	id int,
-	urlOptions map[string]string,
-) (*TVAlternativeTitles, error) {
+func (c *Client) GetTVAlternativeTitles(id int64, urlOptions map[string]string) (*TVAlternativeTitles, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/alternative_titles?api_key=%s%s",
@@ -392,10 +379,7 @@ type TVChanges struct {
 // You can use the and methods to look these up individually.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-changes
-func (c *Client) GetTVChanges(
-	id int,
-	urlOptions map[string]string,
-) (*TVChanges, error) {
+func (c *Client) GetTVChanges(id int64, urlOptions map[string]string) (*TVChanges, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/changes?api_key=%s%s",
@@ -421,10 +405,7 @@ type TVContentRatings struct {
 // GetTVContentRatings get the list of content ratings (certifications) that have been added to a TV show.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-content-ratings
-func (c *Client) GetTVContentRatings(
-	id int,
-	urlOptions map[string]string,
-) (*TVContentRatings, error) {
+func (c *Client) GetTVContentRatings(id int64, urlOptions map[string]string) (*TVContentRatings, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/content_ratings?api_key=%s%s",
@@ -473,10 +454,7 @@ type TVCredits struct {
 // GetTVCredits get the credits (cast and crew) that have been added to a TV show.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-credits
-func (c *Client) GetTVCredits(
-	id int,
-	urlOptions map[string]string,
-) (*TVCredits, error) {
+func (c *Client) GetTVCredits(id int64, urlOptions map[string]string) (*TVCredits, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/credits?api_key=%s%s",
@@ -504,10 +482,7 @@ type TVEpisodeGroups struct {
 // With a group ID you can call the get TV episode group details method.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-episode-groups
-func (c *Client) GetTVEpisodeGroups(
-	id int,
-	urlOptions map[string]string,
-) (*TVEpisodeGroups, error) {
+func (c *Client) GetTVEpisodeGroups(id int64, urlOptions map[string]string) (*TVEpisodeGroups, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/episode_groups?api_key=%s%s",
@@ -548,10 +523,7 @@ type TVExternalIDs struct {
 // *Defunct or no longer available as a service.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-external-ids
-func (c *Client) GetTVExternalIDs(
-	id int,
-	urlOptions map[string]string,
-) (*TVExternalIDs, error) {
+func (c *Client) GetTVExternalIDs(id int64, urlOptions map[string]string) (*TVExternalIDs, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/external_ids?api_key=%s%s",
@@ -599,10 +571,7 @@ type TVImages struct {
 // separated value like so: include_image_language=en,null.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-images
-func (c *Client) GetTVImages(
-	id int,
-	urlOptions map[string]string,
-) (*TVImages, error) {
+func (c *Client) GetTVImages(id int64, urlOptions map[string]string) (*TVImages, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/images?api_key=%s%s",
@@ -628,9 +597,7 @@ type TVKeywords struct {
 // GetTVKeywords get the keywords that have been added to a TV show.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-keywords
-func (c *Client) GetTVKeywords(
-	id int,
-) (*TVKeywords, error) {
+func (c *Client) GetTVKeywords(id int64) (*TVKeywords, error) {
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/keywords?api_key=%s",
 		baseURL,
@@ -656,10 +623,7 @@ type TVRecommendations struct {
 // GetTVRecommendations get the list of TV show recommendations for this item.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-recommendations
-func (c *Client) GetTVRecommendations(
-	id int,
-	urlOptions map[string]string,
-) (*TVRecommendations, error) {
+func (c *Client) GetTVRecommendations(id int64, urlOptions map[string]string) (*TVRecommendations, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/recommendations?api_key=%s%s",
@@ -688,10 +652,7 @@ type TVReviews struct {
 // GetTVReviews get the reviews for a TV show.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-reviews
-func (c *Client) GetTVReviews(
-	id int,
-	urlOptions map[string]string,
-) (*TVReviews, error) {
+func (c *Client) GetTVReviews(id int64, urlOptions map[string]string) (*TVReviews, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/reviews?api_key=%s%s",
@@ -718,9 +679,7 @@ type TVScreenedTheatrically struct {
 // have been screened in a film festival or theatre.
 //
 // https://developers.themoviedb.org/3/tv/get-screened-theatrically
-func (c *Client) GetTVScreenedTheatrically(
-	id int,
-) (*TVScreenedTheatrically, error) {
+func (c *Client) GetTVScreenedTheatrically(id int64) (*TVScreenedTheatrically, error) {
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/screened_theatrically?api_key=%s",
 		baseURL,
@@ -744,10 +703,7 @@ type TVSimilar struct {
 // These items are assembled by looking at keywords and genres.
 //
 // https://developers.themoviedb.org/3/tv/get-similar-tv-shows
-func (c *Client) GetTVSimilar(
-	id int,
-	urlOptions map[string]string,
-) (*TVSimilar, error) {
+func (c *Client) GetTVSimilar(id int64, urlOptions map[string]string) (*TVSimilar, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/similar?api_key=%s%s",
@@ -773,10 +729,7 @@ type TVWatchProviders struct {
 // GetTVWatchProviders get a list of the availabilities per country by provider for a TV show.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-watch-providers
-func (c *Client) GetTVWatchProviders(
-	id int,
-	urlOptions map[string]string,
-) (*TVWatchProviders, error) {
+func (c *Client) GetTVWatchProviders(id int64, urlOptions map[string]string) (*TVWatchProviders, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/watch/providers?api_key=%s%s",
@@ -813,10 +766,7 @@ type TVTranslations struct {
 // GetTVTranslations get a list fo translations that have been created for a TV Show.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-translations
-func (c *Client) GetTVTranslations(
-	id int,
-	urlOptions map[string]string,
-) (*TVTranslations, error) {
+func (c *Client) GetTVTranslations(id int64, urlOptions map[string]string) (*TVTranslations, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/translations?api_key=%s%s",
@@ -842,10 +792,7 @@ type TVVideos struct {
 // GetTVVideos get the videos that have been added to a TV show.
 //
 // https://developers.themoviedb.org/3/tv/get-tv-videos
-func (c *Client) GetTVVideos(
-	id int,
-	urlOptions map[string]string,
-) (*TVVideos, error) {
+func (c *Client) GetTVVideos(id int64, urlOptions map[string]string) (*TVVideos, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/videos?api_key=%s%s",
@@ -1015,11 +962,7 @@ func (c *Client) GetTVTopRated(
 // https://developers.themoviedb.org/3/authentication/how-do-i-generate-a-session-id
 //
 // https://developers.themoviedb.org/3/tv/rate-tv-show
-func (c *Client) PostTVShowRating(
-	id int,
-	rating float32,
-	urlOptions map[string]string,
-) (*Response, error) {
+func (c *Client) PostTVShowRating(id int64, rating float32, urlOptions map[string]string) (*Response, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/rating?api_key=%s&session_id=%s%s",
@@ -1053,10 +996,7 @@ func (c *Client) PostTVShowRating(
 // https://developers.themoviedb.org/3/authentication/how-do-i-generate-a-session-id
 //
 // https://developers.themoviedb.org/3/tv/delete-tv-show-rating
-func (c *Client) DeleteTVShowRating(
-	id int,
-	urlOptions map[string]string,
-) (*Response, error) {
+func (c *Client) DeleteTVShowRating(id int64, urlOptions map[string]string) (*Response, error) {
 	options := c.fmtOptions(urlOptions)
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/rating?api_key=%s&session_id=%s%s",
